Begining the SpellCasting System
There is now a public struct Spell that will hold all relevant information for each spell. I started a SpellCastingSystem object in order to manage the player spell casting process.

In addition, instruments are now set up to be controlled by the music engine. A method name PlayLayeredSegment is now implemented in order to play the variable layed music segments.